refactoring into client and server dirs
