+++ client/src/global.d.ts
