+++ client/src/routes/index.svelte
