+++ client/svelte.config.js
