+++ client/tsconfig.json
@@ -28,5 +28,10 @@
 			"$utils/*": ["src/utils/*"]
 		}
 	},
-	"include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.ts", "src/**/*.svelte"]
+	"include": [
+		"client/src/**/*.d.ts",
+		"client/src/**/*.js",
+		"client/src/**/*.ts",
+		"client/src/**/*.svelte"
+	]
 }
